Добавляет задание по модулям
